+++ eslint.config.js
@@ -7,11 +7,10 @@ export default tseslint.config(
   ...tseslint.configs.recommended,
   {
     rules: {
-      "no-unused-vars": "warn",
-      "@typescript-eslint/no-explicit-any": "warn",
+      "@typescript-eslint/no-explicit-any": "off",
     },
   },
   {
     ignores: ["node_modules/", "lib/", ".gitignore"],
-  },
+  }
 );