+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./defaultHandler.js";
+export * from "./handler.js";
 export * from "./types.js";
-export * from "./routeHandler.js";
+export * from "./error.js";