+++ src/internal/types.ts
@@ -1,6 +1,3 @@
-import { NextApiHandler } from "next";
-import { NextApiHandlerWithError } from "../types.js";
-
 export enum RequestMethod {
   GET = "GET",
   POST = "POST",
@@ -10,15 +7,6 @@ export enum RequestMethod {
 }
 
 export enum ResponseStatus {
-  OK = 200,
-  Unauthorized = 401,
   MethodNotAllowed = 405,
   InternalServerError = 500,
-}
-
-export type Method = keyof typeof RequestMethod;
-
-export interface DefaultConfig {
-  methodFn: NextApiHandler;
-  errorFn: NextApiHandlerWithError;
 }