+++ src/utils.ts
@@ -1,6 +1,0 @@
-export function mergeConfigs<T>(
-  defaultConfig: T,
-  overrideConfig: Partial<T> | undefined
-): T {
-  return { ...defaultConfig, ...overrideConfig };
-}