+++ CHANGELOG.md
@@ -2,8 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog], and this project adheres to
-[Semantic Versioning].
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [1.0.0] - 2024-06-22
+
+### Added
+
+- `DuplicateMethodError` class
+- `NextApiJSONRequest`, `NextApiJSONRequest`, and `NextQueryHandler` utility types
+
+### Changed
+
+- readme documentation
+
+### Removed
+
+- `DefaultHandler` class
+- `Config` for handlers
+- `methodFn` and `errorFn` methods
 
 ## [0.0.1] - 2024-05-01
 
@@ -15,5 +31,4 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 - `NextApiHandlerWithSession` includes an optional generic
 
-[keep a changelog]: https://keepachangelog.com/en/1.0.0/
-[semantic versioning]: https://semver.org/spec/v2.0.0.html
+[keep a changelog]: